Set up basic implementation of the Pokedex class.
Created a static variable to hold the single instance of Pokedex and a
list to store the Pokemon.

Implemented private constructor.

I also added a public static method to get the singleton instance, and
implementation to create a new instance.

Added a getter method to return all pokemon in the pokedex.